test: Suppress warnings

+++ testconv.rb
@@ -103,7 +103,7 @@ static void xx(char* pattern, char* str, int from, int to, int mem, int not)
       }
       else {
         fprintf(stdout, "FAIL: /%s/ '%s' %d-%d : %d-%d\\n", pattern, str,
-	        from, to, pmatch[mem].rm_so, pmatch[mem].rm_eo);
+	        (int)from, (int)to, (int)pmatch[mem].rm_so, (int)pmatch[mem].rm_eo);
         nfail++;
       }
     }
@@ -160,7 +160,7 @@ static void xx(char* pattern, char* str, int from, int to, int mem, int not)
       }
       else {
         fprintf(stdout, "FAIL: /%s/ '%s' %d-%d : %d-%d\\n", pattern, str,
-	        from, to, region->beg[mem], region->end[mem]);
+	        (int)from, (int)to, (int)region->beg[mem], (int)region->end[mem]);
         nfail++;
       }
     }
