+++ testu.c
@@ -106,7 +106,7 @@ static void xx(char* pattern, char* str, int from, int to, int mem, int not)
       }
       else {
         fprintf(stdout, "FAIL: /%s/ '%s' %d-%d : %d-%d\n", cpat, cstr,
-	        from, to, pmatch[mem].rm_so, pmatch[mem].rm_eo);
+	        (int)from, (int)to, (int)pmatch[mem].rm_so, (int)pmatch[mem].rm_eo);
         nfail++;
       }
     }
@@ -181,7 +181,7 @@ static void xx(char* pattern, char* str, int from, int to, int mem, int not)
       }
       else {
         fprintf(stdout, "FAIL: /%s/ '%s' %d-%d : %d-%d\n", cpat, cstr,
-	        from, to, region->beg[mem], region->end[mem]);
+	        (int)from, (int)to, (int)region->beg[mem], (int)region->end[mem]);
         nfail++;
       }
     }
